+++ src/binding/slepc4py/CHANGES.rst
@@ -5,6 +5,12 @@ CHANGES: SLEPc for Python
 :Author:  Lisandro Dalcin
 :Contact: dalcinl@gmail.com
 
+Development
+===========
+
+- In slepc4py now `EPS.getEigenpair()` and `EPS.getEigenvalue()` will return a real value
+instead of a complex, if the problem is of Hermitian or generalized Hermitian type.
+
 
 Release 3.21.0
 ==============